feat: use vscode's implementation to open directory and `openFileExplorer`as a fallback.
Instead of relying on explorer being in the PATH environment variable,
we rely on vscode (which depends on electron) to find the explorer for the specific os.
Hopefully electrons implementation will be more reliable than our own.
We open snippet directory by opening /path/to/folder/snippet.hsnips,
since vscode opens the directory that the file/directory is in instead of opening the directory.
If this doesn't work we try to open the explorer with the command line instead
and if that fails we open the directory that contains the snippet directory.